refactor: Add ComicService for creating and updating comics
Move the mapping from a comic module's metadata to the database out of
comics.core.metadata, which is left responsible for reading metadata and
deciding which comics to read it for.
